backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.2
pydantic-settings==2.6.1
python-multipart==0.0.17
structlog==24.4.0
email-validator==2.2.0
